+++ server.js
@@ -6,6 +6,8 @@ const bodyParser = require('body-parser');
 const generateId = require('./lib/generate-id');
 const Survey = require('./lib/survey');
 const _ = require("lodash");
+const totalVotes = require('./lib/total-votes');
+const votePercentages = require('./lib/vote-percentages');
 const app = express();
 const port = process.env.PORT || 3000;
 const server = http.createServer(app)
@@ -91,19 +93,5 @@ io.on('connection', function (socket) {
   });
 
 });
-
-function votePercentages(responses) {
-  return _.mapValues(responses, function(votes) {
-    return Math.round((votes/totalVotes(responses))*100);
-  });
-}
-
-function totalVotes(responses) {
-  if (_.sum(_.values(responses)) === 0) {
-    return 1;
-  } else {
-    return _.sum(_.values(responses));
-  }
-}
 
 module.exports = server;
